Externalise science module application inside a routine for re-use
