+++ bin/raw2science.py
@@ -27,7 +27,6 @@ from pyspark.sql import functions as F
 
 import argparse
 import time
-import json
 import os
 
 from fink_broker import __version__ as fbvsn
@@ -35,23 +34,12 @@ from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession
 from fink_broker.sparkUtils import connect_to_raw_database
 from fink_broker.filters import apply_user_defined_filter
-from fink_broker.filters import apply_user_defined_processors
 from fink_broker.loggingUtils import get_fink_logger, inspect_application
+
+from fink_broker.science import apply_science_modules
 
 from fink_science import __version__ as fsvsn
 from fink_science import __file__ as fspath
-from fink_science.xmatch.processor import cdsxmatch
-
-from fink_science.random_forest_snia.processor import rfscore_sigmoid_full
-from fink_science.utilities import concat_col
-
-from fink_science.snn.processor import snn_ia
-
-from fink_science.microlensing.processor import mulens
-from fink_science.microlensing.classifier import load_mulens_schema_twobands
-from fink_science.microlensing.classifier import load_external_model
-
-from fink_science.asteroids.processor import roid_catcher
 
 qualitycuts = 'fink_broker.filters.qualitycuts'
 
@@ -71,88 +59,14 @@ def main():
     df = connect_to_raw_database(
         args.rawdatapath, args.rawdatapath + "/*", latestfirst=False)
 
-    # Apply level one filters
+    # Apply quality cuts
     logger.info(qualitycuts)
     df = apply_user_defined_filter(df, qualitycuts)
 
-    # Retrieve time-series information
-    to_expand = [
-        'jd', 'fid', 'magpsf', 'sigmapsf',
-        'magnr', 'sigmagnr', 'magzpsci', 'isdiffpos'
-    ]
+    # Apply science modules
+    df = apply_science_modules(df, logger)
 
-    # Append temp columns with historical + current measurements
-    prefix = 'c'
-    for colname in to_expand:
-        df = concat_col(df, colname, prefix=prefix)
-    expanded = [prefix + i for i in to_expand]
-
-    # Apply level one processor: cdsxmatch
-    logger.info("New processor: cdsxmatch")
-    colnames = [
-        df['objectId'],
-        df['candidate.ra'],
-        df['candidate.dec']
-    ]
-    df = df.withColumn('cdsxmatch', cdsxmatch(*colnames))
-
-    # Apply level one processor: rfscore
-    logger.info("New processor: rfscore")
-
-    # Perform the fit + classification.
-    # Note we can omit the model_path argument, and in that case the
-    # default model `data/models/default-model.obj` will be used.
-    rfscore_args = ['cjd', 'cfid', 'cmagpsf', 'csigmapsf']
-    df = df.withColumn(
-        'rfscore',
-        rfscore_sigmoid_full(*rfscore_args)
-    )
-
-    # Apply level one processor: superNNova
-    logger.info("New processor: supernnova")
-
-    snn_args = ['candid', 'cjd', 'cfid', 'cmagpsf', 'csigmapsf']
-    df = df.withColumn('snnscore', snn_ia(*snn_args))
-
-    # Apply level one processor: microlensing
-    logger.info("New processor: microlensing")
-
-    # broadcast models
-    curdir = os.path.dirname(os.path.abspath(fspath))
-    model_path = curdir + '/data/models/'
-    rf, pca = load_external_model(model_path)
-    rfbcast = spark.sparkContext.broadcast(rf)
-    pcabcast = spark.sparkContext.broadcast(pca)
-
-    def mulens_wrapper(fid, magpsf, sigmapsf, magnr, sigmagnr, magzpsci, isdiffpos):
-        """ Wrapper to mulens pass broadcasted values from this scope
-        """
-        return mulens(
-            fid, magpsf, sigmapsf, magnr,
-            sigmagnr, magzpsci, isdiffpos,
-            rfbcast.value, pcabcast.value)
-
-    # Retrieve schema
-    schema = load_mulens_schema_twobands()
-
-    # Create standard UDF
-    mulens_udf = F.udf(mulens_wrapper, schema)
-
-    # Required alert columns - already computed for SN
-    mulens_args = [
-        'cfid', 'cmagpsf', 'csigmapsf',
-        'cmagnr', 'csigmagnr', 'cmagzpsci', 'cisdiffpos']
-    df = df.withColumn('mulens', mulens_udf(*mulens_args))
-
-    # Apply level one processor: asteroids
-    logger.info("New processor: asteroids")
-    args_roid = ['cfid', 'cmagpsf', 'candidate.ssdistnr']
-    df = df.withColumn('roid', roid_catcher(*args_roid))
-
-    # Drop temp columns
-    df = df.drop(*expanded)
-
-    # Add librarys versions
+    # Add library versions
     df = df.withColumn('fink_broker_version', F.lit(fbvsn))\
         .withColumn('fink_science_version', F.lit(fsvsn))
